DATA COMIC DRAFT
Based on the insights we discussed above, I propose that we connect our insights by telling a story of the virus and CC(Clevercogs) from a first-person perspective.

We can divide the whole period into seven stages. The following is a sketch of the COMIC.
First, I will introduce the background of the story (P1). Before the epidemic, the virus alliance that COVID belongs to wants to invade the earth. Their biggest enemy is CC, which is a robot in the shape of a gear. Here, the main functions, user distribution and click distribution of CC are represented by a pie chart in the shape of a gear.
The first stage is the first National Lockdown (P2-3), where COVID hears of the defeat in Scotland and is ready to start his own battle to defeat CC. Here, a map is used to show the distribution map of the user's city, representing the influence area of CC. COVID comes to the street and finds out that there is no one available in the street in lockdown. The layout is used to show the proportion of clicks taken by the link title. COVID goes to the doorknob of an elderly man who doesn't leave the house because he's so addicted to his game. Cc has added a new function "My Care" as the impact of the pandemic spreads.
The second stage is the minimal lockdown (P4). COVID-19 further affects people's health. People's attention has shifted from entertainment to their own health. At the same time, CC also launched the plan of universal REGISTRATION of GP, which first registered the elderly and then the young. The graph below shows the change of the proportion of link title clicks in this period in the visual way of the planar graph.
In the third stage, Reimposture Restriction (P5), COVID found that arthritis users were very active and decided to target cc for them, for which CC improved the 'My Care' function. This is a podium presentation of the arthritis users who contributed the most clicks.
The fourth phase is the second national lockdown (P6-7), in which COVID-19 was defeated with the help of vaccines and the lockdown, but came back again shortly after. Here is a graph showing the change in the proportion of clicks taken by Link title during this period. The mutated virus continues to endanger people's health. Different diseases have new requirements for nursing during this period. Here is a broken line graph showing the changes in the number of patients with different diseases. CC added the function of "Care Note", which became the first one after its launch. COVID is hard hit.
The fifth stage is Reintroducing (P8). In the confrontation between COVID-19 and CC, the government issued a Tier Four policy, and CC added a 'Responder Service', summoning a bunch of people wearing protective clothing.
Phase 6 is the third National lockdown (P9), in which more and more people have started to realize the seriousness of the pandemic and have started to pay attention to and beat COVID-19. This is a line graph showing the number of searches for COVID-19 related topics by age group.