P4
The second stage is the minimal lockdown (P4). COVID-19 further affects people's health. People's attention has shifted from entertainment to their own health. At the same time, CC also launched the plan of universal REGISTRATION of GP, which first registered the elderly and then the young. The graph below shows the change of the proportion of link title clicks in this period in the visual way of the planar graph.