Create COVID vs CLEVERCOGS.001.jpeg
Before the epidemic, the virus alliance that COVID belongs to wants to invade the earth. Their biggest enemy is CC, which is a robot in the shape of a gear. Here, the main functions, user distribution and click distribution of CC are represented by a pie chart in the shape of a gear.