Create COVID vs CLEVERCOGS.003.jpeg
COVID comes to the street and finds out that there is no one available in the street in lockdown. The layout is used to show the proportion of clicks taken by the link title. COVID goes to the doorknob of an elderly man who doesn't leave the house because he's so addicted to his game. Cc has added a new function "My Care" as the impact of the pandemic spreads.