Create COVID vs CLEVERCOGS.007.jpeg
Here is a graph showing the change in the proportion of clicks taken by Link title during this period. The mutated virus continues to endanger people's health. Different diseases have new requirements for nursing during this period. Here is a broken line graph showing the changes in the number of patients with different diseases. CC added the function of "Care Note", which became the first one after its launch. COVID is hard hit.