Create Data Cleaning Method 1
Data cleaning ideas:

In order to ensure the accuracy of the data, we need to consider the problem of duplicate data. Because the question we asked contains the exploration of age, then we need to think about a certain time and scope.

I selected 2021 years for analysis. First, I found that users had a tendency to perform multiple operations per minute, and we needed to analyze and filter these repeated data.

I use time as the standard, delete multiple repetitions at the same time and keep the last one. Because it helps to determine what function the user ultimately stays on when using this product. This screening can more accurately determine which features are most attractive to users to stay and use. But at the same time, I added a restriction on the visitorID column, because although the probability is very small, there may be two users operating the system at the same time, so only using the similarity in the time column to determine the result is likely to be inaccurate.