Create Data clearing Method 2
part1 ：In the table, the physical health status of the elderly is the same, so there are multiple meaningless counts. First, I need to clear duplicates.

part2: First of all, according to the requirements of the project, we only need to pay attention to the use of users. However, when we checked the Excel table, We found that the people who used the system had many identities. Like staff. So we carried out data cleaning.Next, delete the data row with ExternalID empty.